# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Prospix API
  description: |
    API multi-tenant da plataforma de prospecção inteligente Prospix.
    Primeiro tenant: Giovane Carrara (MetLife · SJRP).

    ## Autenticação
    Todas as rotas (exceto `/auth/*` e `/webhooks/*`) exigem JWT Bearer + header `X-Tenant-Id`.
    O `X-Tenant-Id` é validado contra o JWT — mismatch retorna 403.

    ## Multi-tenancy
    Todo dado é isolado por tenant via RLS no PostgreSQL. Um tenant nunca vê dados de outro.

    ## Rate limits
    - `/auth/*`: 10 req/min por IP
    - `/tenant/*` GET: 600 req/min por tenant
    - `/tenant/*` POST/PATCH: 60 req/min por tenant
  version: 1.0.0
  contact:
    name: Guilds
    email: gustavo.macedo@guilds.com.br
servers:
  - url: https://api.prospix.com.br/v1
    description: Produção
  - url: https://api-staging.prospix.com.br/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Local dev
...
security:
  - bearerAuth: []
    tenantHeader: []
paths:
...
  /tenant/conversations:
    get:
      tags: [Conversations]
      summary: Lista conversas ativas
      x-prospix-response-shape: raw-array
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [active, paused, escalated, closed] } }
        - { name: filter, in: query, schema: { type: string, enum: [hot, waiting, scheduled] } }
        - { name: sort, in: query, schema: { type: string } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        '200':
          description: Lista paginada
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Conversation' }
  /tenant/conversations/{id}/messages:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Conversations]
      summary: Mensagens da conversa (ordem cronológica)
      x-prospix-response-shape: raw-array
      responses:
        '200':
          description: Lista de mensagens
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Message' }
    post:
      tags: [Conversations]
      summary: Envia mensagem (só se user assumiu a conversa)
      x-prospix-response-shape: raw-object
      x-prospix-error-shape: flat-error
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string }
                send_via: { type: string, enum: [whatsapp], default: whatsapp }
      responses:
        '201':
          description: Mensagem enfileirada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '409':
          description: Conversa em modo IA — assuma antes de enviar
  /tenant/conversations/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    patch:
      tags: [Conversations]
      summary: Atualiza controle IA/humano da conversa
      x-prospix-response-shape: raw-object
      x-prospix-error-shape: flat-error
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [aiHandling]
              properties:
                aiHandling: { type: boolean }
      responses:
        '200':
          description: Conversa atualizada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Conversation' }
        '400': { description: Payload invalido }
        '404': { $ref: '#/components/responses/NotFound' }

  # ===========================================================================
  # MEETINGS
  # ===========================================================================
...
  /tenant/scripts:
    get:
      tags: [Scripts]
      summary: Lista roteiros do tenant
      x-prospix-response-shape: raw-array
      responses:
        '200':
          description: Lista
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Script' }
    post:
      tags: [Scripts]
      summary: Cria roteiro do tenant
      x-prospix-response-shape: raw-object
      x-prospix-error-shape: flat-error
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [baseMessage]
              properties:
                name: { type: string }
                baseMessage: { type: string }
                variations:
                  type: array
                  items:
                    type: object
                    properties:
                      id: { type: string }
                      name: { type: string }
                      weight: { type: number, minimum: 0, maximum: 100 }
                      content: { type: string }
      responses:
        '201':
          description: Script criado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Script' }
        '400': { description: Payload invalido }
  /tenant/scripts/simulate:
    post:
      tags: [Scripts]
      summary: Simula resposta de roteiro
      x-prospix-response-shape: raw-object
      x-prospix-error-shape: flat-error
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [input, baseMessage]
              properties:
                input: { type: string }
                baseMessage: { type: string }
                variations:
                  type: array
                  items:
                    type: object
                    properties:
                      name: { type: string }
                      weight: { type: number, minimum: 0, maximum: 100 }
                      content: { type: string }
      responses:
        '200':
          description: Preview simulado
          content:
            application/json:
              schema:
                type: object
                properties:
                  reply: { type: string }
                  variantUsed: { type: string }
        '400': { description: Payload invalido }
  /tenant/scripts/clone:
    post:
      tags: [Scripts]
      summary: Clona template pra tenant
      x-prospix-response-shape: raw-object
      x-prospix-error-shape: flat-error
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [templateId]
              properties:
                templateId: { type: string, format: uuid }
      responses:
        '201':
          description: Script clonado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Script' }
        '400': { description: Payload invalido }
  /tenant/scripts/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    patch:
      tags: [Scripts]
      summary: Atualiza roteiro (flow, status)
      x-prospix-response-shape: raw-object
      x-prospix-error-shape: flat-error
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                flow: { type: object }
                status: { type: string, enum: [DRAFT, ACTIVE, PAUSED, ARCHIVED] }
                baseMessage: { type: string }
      responses:
        '200':
          description: Atualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Script' }
        '400': { description: Payload invalido }
  /tenant/scripts/{id}/variations:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [Scripts]
      summary: Cria ou atualiza variante de roteiro
      x-prospix-response-shape: raw-object
      x-prospix-error-shape: flat-error
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [variantLetter, message]
              properties:
                variantLetter: { type: string, minLength: 1, maxLength: 2 }
                message: { type: string }
                weight: { type: number, minimum: 0, maximum: 1, default: 0.33 }
      responses:
        '201':
          description: Variante salva
          content:
            application/json:
              schema: { type: object }
        '400': { description: Payload invalido }
        '404': { $ref: '#/components/responses/NotFound' }
  /tenant/scripts/{id}/test:
    post:
      tags: [Scripts]
      summary: Testa roteiro com lead fictício
      x-prospix-response-shape: raw-object
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                test_lead: { type: object }
      responses:
        '200':
          description: Preview
          content:
            application/json:
              schema:
                type: object
                properties:
                  preview: { type: string }
                  nodes_count: { type: integer }
                  active_variations:
                    type: array
                    items: { type: string }

  # ===========================================================================
  # DASHBOARD
  # ===========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    tenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-Id
  responses:
...
    NotFound:
      description: Recurso não encontrado neste tenant
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum: [UNAUTHENTICATED, UNAUTHORIZED, RESOURCE_NOT_FOUND, VALIDATION_ERROR, RATE_LIMITED, TENANT_QUOTA_EXCEEDED, EXTERNAL_SERVICE_DOWN, INTERNAL_ERROR]
            message: { type: string }
            details: { type: object }
            trace_id: { type: string }
        meta:
          type: object
          properties:
            request_id: { type: string }
            timestamp: { type: string, format: date-time }
...
    Profession:
      type: string
      enum: [doctor, lawyer, dentist, entrepreneur, engineer, architect, accountant, other]
...
    Conversation:
      type: object
      properties:
        id: { type: string, format: uuid }
        lead_id: { type: string, format: uuid }
        lead_name: { type: string }
        status: { type: string, enum: [active, paused, escalated, closed] }
        ai_handling: { type: boolean }
        message_count: { type: integer }
        last_message_at: { type: string, format: date-time }
        last_message_preview: { type: string }
    Message:
      type: object
      properties:
        id: { type: string, format: uuid }
        direction: { type: string, enum: [inbound, outbound] }
        sender: { type: string, enum: [ai, user, lead] }
        content: { type: string }
        delivery_status: { type: string, enum: [queued, sent, delivered, read, failed] }
        intent_detected: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
...
    Script:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        category: { type: string, enum: [approach, objection, education, closing, follow_up, referral, reactivation] }
        target_profession: { $ref: '#/components/schemas/Profession' }
        status: { type: string, enum: [draft, active, archived] }
        response_rate: { type: number }
        conversion_rate: { type: number }
        total_usages: { type: integer }
...